+++ CHANGELOG.md
@@ -24,7 +24,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `fetch_dataframe` utility function
 - Added data assimilation model class
 - Added equirectangular interpolation data assimilation model
-- Adding Beta serve utils with inference server and client implementations
+- Added StormCast SDA model
+- Added beta serve utils with inference server and client implementations
+- Added HealPix data assimilation (HealDA) model
 - Added `energy_score` metric for multivariate ensemble forecast verification
 
 ### Changed
@@ -33,6 +35,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed ISD data frame return to master schema
 - handshake_coords is now accepting list of dimensions while remaining backwards-compatible
 - Updated CBottle infill to mixture of model checkpoints
+- Updated GraphCastOperational and GraphCastSmall latitude input / output to be [90,-90]
+- Updated GraphCast models to support multiple time inputs, multiple times will be looped not batched
+- Renamed `tolerance` parameter in ISD data source to `time_tolerance`
 
 ### Deprecated
 
@@ -42,8 +47,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fixed timezone bug in `CBottleVideo` that converted UTC time to local system time
 - Bug in cbottle datasource resulting identical samples for multiple samples
 - Bug in StormCast loading out-of-date model package, introduced in `5518edecbabee371c824b34f0f2ec269a4d6094f`
+- Bug in spherical perturbations which did not use lmax from the SHT transform
 
 ### Security